requirements.txt
Django
django-multiselectfield
python-dotenv
gunicorn
